requirements.txt
Flask
pdfminer.six
requests
gunicorn
